+++ src/codegen/unwinding.cpp
@@ -1219,6 +1219,95 @@ void logByCurrentPythonLine(const std::string& stat_name) {
     Stats::log(Stats::getStatCounter(stat));
 }
 
+void _printStacktrace() {
+    static bool recursive = false;
+
+    if (recursive) {
+        fprintf(stderr, "_printStacktrace ran into an issue; refusing to try it again!\n");
+        return;
+    }
+
+    recursive = true;
+    printTraceback(getTraceback());
+    recursive = false;
+}
+
+extern "C" void abort() {
+    static void (*libc_abort)() = (void (*)())dlsym(RTLD_NEXT, "abort");
+
+    // In case displaying the traceback recursively calls abort:
+    static bool recursive = false;
+
+    if (!recursive) {
+        recursive = true;
+        Stats::dump();
+        fprintf(stderr, "Someone called abort!\n");
+
+        // If traceback_cls is NULL, then we somehow died early on, and won't be able to display a traceback.
+        if (traceback_cls) {
+
+            // If we call abort(), things may be seriously wrong.  Set an alarm() to
+            // try to handle cases that we would just hang.
+            // (Ex if we abort() from a static constructor, and _printStackTrace uses
+            // that object, _printStackTrace will hang waiting for the first construction
+            // to finish.)
+            alarm(1);
+            try {
+                _printStacktrace();
+            } catch (ExcInfo) {
+                fprintf(stderr, "error printing stack trace during abort()");
+            }
+
+            // Cancel the alarm.
+            // This is helpful for when running in a debugger, since otherwise the debugger will catch the
+            // abort and let you investigate, but the alarm will still come back to kill the program.
+            alarm(0);
+        }
+    }
+
+    if (PAUSE_AT_ABORT) {
+        fprintf(stderr, "PID %d about to call libc abort; pausing for a debugger...\n", getpid());
+
+        // Sometimes stderr isn't available (or doesn't immediately appear), so write out a file
+        // just in case:
+        FILE* f = fopen("pausing.txt", "w");
+        if (f) {
+            fprintf(f, "PID %d about to call libc abort; pausing for a debugger...\n", getpid());
+            fclose(f);
+        }
+
+        while (true) {
+            sleep(1);
+        }
+    }
+    libc_abort();
+    __builtin_unreachable();
+}
+
+#if 0
+extern "C" void exit(int code) {
+    static void (*libc_exit)(int) = (void (*)(int))dlsym(RTLD_NEXT, "exit");
+
+    if (code == 0) {
+        libc_exit(0);
+        __builtin_unreachable();
+    }
+
+    fprintf(stderr, "Someone called exit with code=%d!\n", code);
+
+    // In case something calls exit down the line:
+    static bool recursive = false;
+    if (!recursive) {
+        recursive = true;
+
+        _printStacktrace();
+    }
+
+    libc_exit(code);
+    __builtin_unreachable();
+}
+#endif
+
 llvm::JITEventListener* makeTracebacksListener() {
     return new TracebacksEventListener();
 }
